Converted client keep alive packet
